feat: add physics ID UI component
This time I will do a split because physics has a lot to add. Oh also, I won't commit this because, um do I really have to explain it? But basically because it's not CreateaAnims functionality, but it was convenient to use CreateAnims code to figure it out. You know, the unused physics. I already created the physics files with just $80 in case of being unused. And they can be used, but after I do the Edit Physics part.