feat: add value edit functions for UndoRedo
I was originally doing the disable enable, but I really couldn't since the stack was being corrupted with these kinda calls. So I'm first doing this commit, and next commit I'm adding the buttons. Also yes I'm aware that the testing was possible with fewer changes but come on. Now it's a complete commit.