feat: undo redo for ColorPickerRectangle
And here's the thing. Yes I'm continuing my explanation of previous commit. Because that value wasn't being set, this feature wasn't working. So you could see it as, well that's necessary to make this feature work, so there is no bug, there is a change necessary to make a feature work. But really, it's all about perspective, it works now so. But I still like to make these kind of analysis to better understand what's going on and what I'm doing with my life basically. Does it change anything? I think so.