feat: UndoRedo when opening and closing Physics Window (via Ctrl-Z)
I mean via Undo. The rest I handle via different commits, because it has a lot of complexity.

Oh! Amend, I want to add, super important that undo and redo happen at the very and I mean very end because of the wait_window. They will be stuck here until the window closes. It doesn't really matter, just as long as they ran last, so that the stack is already updated and the coming Undo and Redo don't break. It's awesome.