feat: add UndoRedo for removal and insertion of physics columns
So, so awesome. Many tricks here like updating self.frame_index logic for insertion such that it doesn't add the +2, we add it in the PhysicsLabel.insert, and then playing with the numbers such that after an insertion, removing at the same position will leave it all exactly as it was. So, so awesome.