feat: add UndoRedo for closing Physics Window using the X button
So yes, before this commit, if you used the X button, doing Ctrl-Z and stuff could really break stuff. But now it works awesomely well, and awesomely well with switch UndoRedo branch too! I was going to add it to its own commit but they're closely related. So here it is! So awesome!