refactor: remove amount_changes and use trace
I left explanation and details in comment but amount_changes, it can be done yes but it's harder to read, when you're at say -2 and then you do new changes, you need to add a lot of logic such that for example now a new movement or a new transaction I guess you could say, it goes to -3 instead of -1, so stuff like that. I prefer now to use the trace, it seems more intuitive to me now.