fix: when saving CHR palette with All Files and no extension it saves with .pal instead of .chr.pal
Yes, here again I'm describing the issue from user experience perspective. Internally, the bug is the defaultextension which should be .chr.pal to be consistent with the rest of the fields, but one could say hey ok that's strange but no impact. But yes, there is impact and it's this so, yeah just updated. Sometimes we don't know but just in case might be great to update, here we know so I'm saying it.