feat: update check for updates logic to check for latest release instead
Not a refactor. The thing is, it might never happen in practice but it doesn't leave me sleeping at ease thinking that someone might check for an update in-between I pushed the updates, but I haven't created the release yet! So what will happen is a 404 HTTP error because yes, the version has been updated in the repo but the zip is not ready yet. Those are two different things. It's the pattern or situation where you have two similar things but they're still different and you're checking on the first instead of the second. So, we were checking the first, we will check the second now. To be able to do the download, the release must have been created. Also, another advantage of this method is that it is even faster because it's only one call, yay!